+++ modules/os2web_audit_entity/src/Form/SettingsForm.php
@@ -3,6 +3,7 @@
 namespace Drupal\os2web_audit_entity\Form;
 
 use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\Core\Config\TypedConfigManagerInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\ConfigFormBase;
 use Drupal\Core\Form\FormStateInterface;
@@ -20,9 +21,10 @@ class SettingsForm extends ConfigFormBase {
    */
   public function __construct(
     ConfigFactoryInterface $configFactory,
-    private EntityTypeManagerInterface $entityTypeManager,
+    TypedConfigManagerInterface $typedConfigManager,
+    private EntityTypeManagerInterface $entityTypeManager
   ) {
-    parent::__construct($configFactory);
+    parent::__construct($configFactory, $typedConfigManager);
   }
 
   /**
@@ -31,6 +33,7 @@ class SettingsForm extends ConfigFormBase {
   public static function create(ContainerInterface $container): static {
     return new static(
       $container->get('config.factory'),
+      $container->get('config.typed'),
       $container->get('entity_type.manager')
     );
   }