+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fix Drupal 11 compatibility - settings forms
+
 ## [1.2.0] - 2025-07-03
 
 - Added drupal 11 compatibility